requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Failsafe==0.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
memoization==0.1.4
Pillow==6.1.0
pydarksky==2.0.1
Pyphen==0.9.5
requests==2.22.0
urllib3==1.25.3
Werkzeug==0.15.5
